libs: map: move utils-map into dedicated folder
Relocates utils-map.ts under src/libs/map/ to group all map-related
helpers together. Import paths in PoiMapArrows, the Map widget and
MissionPlanningView are updated to the new location; no behavior change.
